requirements.txt
asgiref==3.8.1
astroid==3.2.2
dill==0.3.8
Django==5.0.6
django-debug-toolbar==4.4.5
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==1.7.1
pylint==3.2.5
python-dotenv==1.0.1
sqlparse==0.5.0
tomlkit==0.12.5
